Creates also a static solution for the scenario.
The "static" (in the lack of imagination for a better name) solution
does DLL references (not project references) to all sub projects. Once
the usual solution is built this one can be used to go straight to the
memory exception without having to build the whole scenario again.

Bridge always runs the project build even if the assembly is up-to-date.